add current user api

+++ api/generate.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
 
+set -eu
+
 cd $(dirname $0)/.. # move to project root
 
 destdir=frontend/src/api
-rm -r $destdir
+test -d $destdir && rm -r $destdir && mkdir -p $destdir
 docker run --rm -u "$(id -u):$(id -g)" -v $PWD:/local openapitools/openapi-generator-cli:v4.3.1 \
   generate \
   -g typescript-fetch \
